backend/requirements_ugc.txt
# Dependencias para sistema UGC
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.1.1
email-validator==2.1.0
python-multipart==0.0.6

